% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVOT2emuDB.R
\name{addVOT2emuDB}
\alias{addVOT2emuDB}
\title{Predict voice onset time and add results to EMU database}
\usage{
addVOT2emuDB(
  emuDB,
  seg_list,
  level = NULL,
  new_level_name = "vot",
  verbose = TRUE,
  sign = c("positive", "negative"),
  pos_params_list = NULL,
  neg_params_list = NULL
)
}
\arguments{
\item{emuDB}{The handle of an EMU database which is already loaded in R.}

\item{seg_list}{A segment list returned by \link[emuR:query]{emuR::query} containing stops
of interest. Alternatively one or more strings giving the characters used to
annotate stops in \code{level}. In this case, does not have to be a perfect match;
the function will
search for all intervals where the first character matches one of the strings
in \code{seg_list}. If any intervals contain the IPA stress symbol, this symbol
will be ignored. \link[emuR:query]{emuR::query} is then called under the hood.}

\item{level}{A string giving the name of the annotation level in \code{emuDB}
to search for instances of segments given in \code{seg_list}. Default is \code{NULL};
not used if a \code{seg_list} returned by \link[emuR:query]{emuR::query} is provided.}

\item{new_level_name}{A string given the name of the new annotation level
used to store predicted voice onset time. Default is \code{vot}.}

\item{verbose}{Logical; should messages be printed in the console indicating
the progress? Default is \code{TRUE}.}

\item{sign}{A string giving the type of voice onset time to predict;
legal values are \code{positive}, \code{negative}, or both. Default is both, in which
case \link{getVOT} tries to predict whether voice onset time is positive or
negative. The development of this function is still in the early stages,
so it is recommended to specify whether VOT is expected to be positive or
negative when possible.}

\item{pos_params_list}{Named list of parameters used to predict positive
voice onset time; see \link{positiveVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}

\item{neg_params_list}{Named list of parameters used to predict negative
voice onset time; see \link{negativeVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}
}
\description{
Predicts voice onset time for all intervals in an EMU database that
match a pattern provided by the user. Adds an additional level definition
to all annotations in the EMU database giving the location of predicted
voice onset time (i.e. pre-voicing or stop release).
}
\examples{
#later
}
\seealso{
This function is intended for users of the EMU-SDMS database system
and the \link{emuR} package. Sister functions which are based on the acoustic
processing software Praat are also available in \link{addVOT2TG} and \link{VOT2newTG},
and these are significantly faster and less computationally expensive; if
possible, it is recommended to use these instead.
Voice onset time is predicted using the \link{getVOT} function, which
calls either the \link{positiveVOT} or \link{negativeVOT} function (or both) for a
given token.
}
