% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVOT2TG.R
\name{addVOT2TG}
\alias{addVOT2TG}
\title{Predict voice onset time and add results to existing TextGrid}
\usage{
addVOT2TG(
  directory,
  tg_tier,
  seg_list,
  verbose = TRUE,
  new_tier_name = "vot",
  sign = c("positive", "negative"),
  pos_params_list = NULL,
  neg_params_list = NULL,
  vo_only = FALSE,
  rel_offset = 0
)
}
\arguments{
\item{directory}{A string giving the location of a directory with pairs of
WAV files and TextGrids.
The results will be stored in a new subdirectory called \code{tg}.}

\item{tg_tier}{String giving the name of a tier in the TextGrids containing
intervals with stops. Alternatively an integer giving the index of a tier
containing intervals with stops stops.}

\item{seg_list}{One or more strings giving the characters used to annotate
stops in \code{tg_tier}. Does not have to be a perfect match; the function will
search for all intervals where the first character matches one of the strings
in \code{seg_list}. If any intervals contain the IPA stress symbol, this symbol
will be ignored.}

\item{verbose}{Logical; should messages be printed in the console indicating
the progress? Default is \code{TRUE}.}

\item{new_tier_name}{String giving the name of the new tier containing
VOT predictions. Default is \code{vot}.}

\item{sign}{A string giving the type of voice onset time to predict;
legal values are \code{positive}, \code{negative}, or both. Default is both, in which
case \link{getVOT} tries to predict whether voice onset time is positive or
negative. The development of this function is still in the early stages,
so it is recommended to specify whether VOT is expected to be positive or
negative when possible.}

\item{pos_params_list}{Named list of parameters used to predict positive
voice onset time; see \link{positiveVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}

\item{neg_params_list}{Named list of parameters used to predict negative
voice onset time; see \link{negativeVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}

\item{vo_only}{Boolean; default is \code{FALSE}. Can be set to \code{TRUE} if
\code{sign='positive'}, and the data is already aligned such that
intervals are aligned to the release. In this case, the burst location is
set as the beginning of the interval, and only voicing onset location is
predicted.}

\item{rel_offset}{Numeric, default is \code{0}. If \code{vo_only=TRUE}, the algorithm
may perform poorly if there is periodicity from e.g. voicing bleed early on
in the interval (this is especially likely if used to delimit fricatives).
\code{rel_offset} tells \code{getVOT} how much of the initial portion of an interval
to ignore when looking for voicing onset (in seconds).}
}
\description{
Loops through a directory with pairs of WAV files and Praat TextGrids,
and predicts voice onset time for all intervals that match a pattern
provided by the user. Copies of the TextGrid files are generated with an
additional tier giving the location of the predicted voice onset time
(i.e. pre-voicing or stop release).
}
\examples{
datapath <- system.file('extdata/add2tg', package='getVOT')
addVOT2TG(directory=datapath, tg_tier='KAN-MAU', seg_list='b')
}
\seealso{
WAV files are imported and TextGrids are exported using the \code{rPraat}
package. Voice onset time is predicted using the \link{getVOT} function, which
calls either the \link{positiveVOT} or \link{negativeVOT} function (or both) for a
given token. See also the sister function \link{VOT2newTG} which generates new
TextGrids with predicted voice onset time for sound files that consist of
a single stop-initial syllable or word.
}
