% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closure.R
\name{find_closure}
\alias{find_closure}
\title{Predict location of stop closure}
\usage{
find_closure(sound, sr, closure_interval = 10)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{sr}{Integer; sample rate of \code{sound}.}

\item{closure_interval}{Numeric. \code{closure_interval} determines the duration
of silent intervals to look for in ms. Default is \code{10}.}
}
\value{
A numeric value giving the sample number at the centre of the
predicted stop closure interval.
}
\description{
Predicts the location of a stop closure in a short sound snippet by
chunking up the first half of the snippet into equidistant intervals and
finding the overall most silent interval.
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs

clo <- find_closure(sig, sr)
}
\seealso{
This function is typically called by either \link{positiveVOT} or
\link{negativeVOT}, which are in turn typically called when using \link{VOT2newTG}
or \link{addVOT2TG} to add predicted VOT to a Praat TextGrid or to an
EMU database with \link{addVOT2emuDB}.
}
