% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVOT.R
\name{getVOT}
\alias{getVOT}
\title{Predict voice onset time}
\usage{
getVOT(
  sound,
  sr,
  sign = c("positive", "negative"),
  neg_params_list = NULL,
  pos_params_list = NULL
)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{sr}{Integer; sample rate of \code{sound}.}

\item{sign}{A string giving the type of voice onset time to predict;
legal values are \code{positive}, \code{negative}, or both. Default is both, in which
case the function tries to predict whether voice onset time is positive or
negative. The development of this function is still in the early stages,
so it is recommended to specify whether VOT is expected to be positive or
negative when possible.}

\item{neg_params_list}{Named list of parameters used to predict negative
voice onset time; see \link{negativeVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}

\item{pos_params_list}{Named list of parameters used to predict positive
voice onset time; see \link{positiveVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}
}
\value{
A named list with the following elements:
\itemize{
\item \code{vo} Gives the sample number where the onset of voicing is predicted.
\item \code{rel} Gives the sample number where the stop burst is predicted.
\item \code{vot} Gives the predicted voice onset time in ms.
}
}
\description{
Voice onset time and the location of the associated acoustic landmarks are
predicted for a sound snippet including a stop consonant.
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs

getVOT(sound=sig, sr=sr)
}
\seealso{
This function calls either \link{positiveVOT}, \link{negativeVOT} or both.
It is used under the hood for adding predicted VOT to a Praat TextGrid with
\link{VOT2newTG} or \link{addVOT2TG} or to an EMU database with \link{addVOT2emuDB}.
}
