% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeVOT.R
\name{negativeVOT}
\alias{negativeVOT}
\title{Predict negative voice onset time and plot wave with segmented pre-voicing}
\usage{
negativeVOT(
  sound,
  sr,
  vo_method = "acf",
  closure_interval = 10,
  vo_granularity = 1.2,
  vo_param = 0.9,
  f0_wl = 50,
  f0_minacf = 0.5,
  vo_only = FALSE,
  rel_method = "transient",
  plot = TRUE,
  params_list = NULL
)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{sr}{Integer; sample rate of \code{sound}.}

\item{vo_method}{String giving the method for predicting the onset of voicing.
There are two legal options:
\itemize{
\item \code{acf} (default). The onset of voicing is predicted from the mean degree
of energy autocorrelation in short intervals of \code{sound} relative to the most
autocorrelated interval during the vowel (see \code{vo_param}). If audio quality
is sufficiently high and stable, \code{acf} usually gives the best results.
\item \code{f0}. The onset of voicing is predicted using the pitch tracking
algorithm implemented in \link[phonTools:pitchtrack]{phonTools::pitchtrack}.
}}

\item{closure_interval}{Numeric. The rough location of the voicing onset is
predicted by first finding the quietest interval in the first half of \code{sound};
\code{closure_interval} determines the duration of intervals to look for in ms.
Default is \code{10}.}

\item{vo_granularity}{Numeric, only used when \code{vo_method='acf'}. Mean energy
autocorrelation is calculated for intervals of this duration (in ms). Default
is \code{1.2}.}

\item{vo_param}{Numeric, only used when \code{vo_method='acf'}. Voicing onset
is predicted by comparing the mean autocorrelation in intervals of size
\code{vo_granularity} after the predicted closure
to the interval with the highest autocorrelation,
and finding the first interval where mean autocorrelation is at this
proportion of the most autocorrelated interval. Default is \code{0.9}.}

\item{f0_wl}{Numeric, only used when \code{vo_method='f0'}. The length of the
analysis window in ms passed on to \link[phonTools:pitchtrack]{phonTools::pitchtrack}. Default is \code{50}.}

\item{f0_minacf}{Numeric, only used when \code{vo_method='f0'}. Autocorrelation
values below this are ignored by \link[phonTools:pitchtrack]{phonTools::pitchtrack}. Default is \code{0.5}.}

\item{vo_only}{Logical; if \code{TRUE}, only voicing onset is predicted.}

\item{rel_method}{String giving the method for predicting the location of the
burst. There are two legal options:
\itemize{
\item \code{transient} (default). The location of the burst is predicted by searching
for the transient phase. \link[phonTools:spectralslice]{phonTools::spectralslice} is used to generate
FFT spectra of short portions of \code{sound} after the predicted onset of voicing,
and the smoothest spectrum is predicted to be the location of the transient.
If audio quality is sufficiently high and stable, \code{transient} usually gives
the best results.
\item \code{amplitude}. The location of the burst is predicted by searching for a
sudden increase in amplitude (which usually signifies the beginning of the
vowel, which in pre-voiced stops should be very close to the burst).
In practice, this is done by recording the highest amplitude in up to 500
1ms chunks of \code{sound} after the predicted voicing onset, smoothing the
resulting time series using a discrete cosine transformation using
\link[emuR:dct]{emuR::dct}, and locating the portion of the resulting function with the
highest velocity.
}}

\item{plot}{Logical; should the results be plotted? Default is \code{TRUE}.}

\item{params_list}{A named list containing the above parameters for
determining the onset of voicing. Usually returned by \code{\link[=neg_setParams]{neg_setParams()}}, but
can also be written by hand. Should have the following named elements (order
not important; see details above):
\itemize{
\item \code{closure_interval}
\item \code{vo_granularity}
\item \code{vo_param}
\item \code{vo_method}
\item \code{f0_wl}
\item \code{f0_minacf}
\item \code{rel_method}
If a \code{params_list} is provided, it will override any corresponding
parameters in the function call.
}}
}
\value{
A named list with the following elements:
\itemize{
\item \code{vo} Gives the sample number where the onset of voicing is predicted.
\item \code{rel} Gives the sample number where the stop burst is predicted.
\item \code{vot} Gives the predicted voice onset time in ms.
\item \code{voi_int} Gives a number with the predicted duration of voicing in
\code{sound} in ms. Potentially used by the \link{getVOT} function to predict whether
voice onset time is positive or negative.
}
}
\description{
Negative voice onset time and the location of pre-voicing is predicted for a
sound snippet containing a pre-voiced stop. Optionally, the wave is plotted
showing the predicted location of pre-voicing.
}
\examples{
datapath <- system.file('extdata/vd', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs

#Default settings do not give very good results for this token
negativeVOT(sound=sig, sr=sr)

#Here with some optimized parameters
negativeVOT(sound=sig, sr=sr, closure_interval=3, vo_granularity=1.55,
vo_param=0.73)
}
\seealso{
This function is called by the more general and somewhat less
flexible function \link{getVOT}, which is in turn called by the functions
\link{VOT2newTG}, \link{addVOT2TG}, and \link{addVOT2emuDB} for bulk analyzing sounds
and annotating the output in new or existing TextGrids (for users of the
acoustic analysis software Praat) or an EMU database (for users of \link{emuR}).
The parameters can be be somewhat obscure and the default parameters do not
necessarily scale particularly well. If a small annotated data set
is provided, \link{neg_setParams} can be used to tune the parameters to your data
and annotation style.
}
