% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_training_diff.R
\name{plot_training_diff}
\alias{plot_training_diff}
\title{Plot hand-annotated voice onset time and predicted voice onset time together}
\usage{
plot_training_diff(directory, params_list, cat = FALSE)
}
\arguments{
\item{directory}{A string giving the location of a directory with WAV/TextGrid
pairs, previously used to optimize parameters with \link{neg_setParams} or
\link{pos_setParams}.}

\item{params_list}{A named list of optimized parameters returned by
\link{neg_setParams} or \link{pos_setParams}.}

\item{cat}{Logical; default is \code{FALSE}. If there are empty frames in the
resulting plot, should they be filled with pictures of my very nice cat
Kipawsky?}
}
\description{
Used to plot hand-annotated voice onset time ("training data") together with
voice onset time that has been predicted using optimized parameters based on
that training data. This is practical in order to check the performance of
the selected parameters. Predicted voice onset time is shown in red, and
hand-annotated voice onset time is shown in blue.
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
p <- neg_setParams(directory=datapath, plot=FALSE)
plot_training_diff(directory=datapath, params_list=p)
}
