% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_setParams.R
\name{pos_setParams}
\alias{pos_setParams}
\title{Optimize parameters for predicting positive voice onset time}
\usage{
pos_setParams(
  directory,
  when_use_f0 = 10,
  when_use_f0_first = 10,
  verbose = TRUE,
  plot = TRUE
)
}
\arguments{
\item{directory}{A string giving the location of a directory with WAV/TextGrid
pairs. Ideally, each WAV file should contain just one syllable with a stop,
where the stop closure is placed somewhere in the first half of the sound file.
The TextGrid tier containing hand-segmented VOT should be named \code{vot}.}

\item{when_use_f0}{Numeric; default is \code{10}. Per default, voicing onset is
predicted with \link{positiveVOT} using \code{method='acf'}. This is the preferred method
because it is significantly faster than the alternative \code{method='f0'}. For
this reason, parameters with \code{method='f0'} are only tested if the average
prediction error with \code{method='acf'} is above a certain threshold.
\code{when_use_f0} sets this threshold; i.e. by default, parameters for
\code{method='f0'} are tested when the average prediction error with \code{method='acf'}
is above 10 ms.}

\item{when_use_f0_first}{Numeric; default is \code{10}. Per default, the burst is
located by first predicting the rough location of the stop closure (i.e.
\link{positiveVOT} is called with \code{f0_first=FALSE}). This is the preferred method
because it is significantly faster than the alternative method, where the
burst is predicted by first finding the location of the vowel and moving
backwards (\code{f0_first=TRUE}). For this reason, parameters with \code{f0_first=TRUE}
are only tested if the average prediction error with \code{f0_first=FALSE} is
above a certain threshold. \code{when_use_f0_first} sets this threshold; by
default, parameters with \code{f0_first=TRUE} are tested when prediction error
for half of the training material is above 10 ms.}

\item{verbose}{Logical; should messages be printed in the console indicating
the progress? Default is \code{TRUE}.}

\item{plot}{Logical; should predicted VOT using the resulting parameters
be plotted against the hand-annotated VOT?}
}
\value{
A named list with optimized parameters for the training data; see
\link{positiveVOT} for more information.
}
\description{
If the user provides a few examples of WAV/TextGrid pairs with hand-annotated
voice onset time of stops, this function will attempt to predict VOT with
a large range of parameters, and select those parameters that on average
minimize prediction error the most. These parameters can then be used for
the rest of the data set.
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
pos_setParams(directory=datapath)
}
\seealso{
The returned list can be passed to the \link{positiveVOT} argument
\code{params_list}, or the argument \code{pos_params_list} in other functions
\link{getVOT}, \link{VOT2newTG}, \link{addVOT2TG}, \link{addVOT2emuDB}. \link{neg_setParams} is a
sister function for optimizing parameters for negative positive voice
onset time. \link{plot_training_diff} can be used to illustrate how closely the
optimized parameters predict the hand-annotated training data.
}
