% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positiveVOT.R
\name{positiveVOT}
\alias{positiveVOT}
\title{Predict positive voice onset time and plot wave with segmented stop release}
\usage{
positiveVOT(
  sound,
  sr,
  closure_interval = 10,
  release_param = 15,
  rel_method = "diff",
  vo_method = "acf",
  vo_granularity = 1,
  vo_param = 0.85,
  f0_wl = 30,
  f0_minacf = 0.5,
  zcr_min = 0.05,
  soe_min = 0.01,
  burst_only = FALSE,
  vo_only = FALSE,
  rel_offset = 0,
  f0_first = FALSE,
  plot = TRUE,
  params_list = NULL
)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{sr}{Integer; sample rate of \code{sound}.}

\item{closure_interval}{Numeric. The rough location of the stop burst is
predicted by first finding the quietest interval in the first half of \code{sound};
\code{closure_interval} determines the duration of intervals to look for in ms.
Default is \code{10}.}

\item{release_param}{Numeric. The location of the burst is predicted by
searching for the first instance after the
estimated closure where a 1 ms interval has
an energy equal to or higher than this proportion of the maximum energy in
\code{sound}. Default is \code{15}.}

\item{rel_method}{String giving the method for predicting the location of the
burst. There are two legal options:
\itemize{
\item \code{raw}. The location of the burst is predicted by searching for
the first instance in the waveform where the amplitude is above the value
set by \code{release_param}.
\item \code{diff} (default). The procedure is the same as in \code{raw}, but the signal is
differenced first.
}}

\item{vo_method}{String giving the method for predicting the onset of voicing.
There are two legal options:
\itemize{
\item \code{acf} (default). The onset of voicing is predicted from the mean degree
of energy autocorrelation in short intervals of \code{sound} relative to the most
autocorrelated interval during the vowel (see \code{vo_param}). If audio quality
is sufficiently high and stable, \code{acf} usually gives the best results.
See \link{vo_acf}.
\item \code{f0}. The onset of voicing is predicted using the pitch tracking
algorithm implemented in \link[phonTools:pitchtrack]{phonTools::pitchtrack}.
\item \code{zcr}. The onset of voicing is predicted from the zero crossing rate.
ZCR is calculated in
5 ms windows with 80\% overlap, and the resulting time series is smoothed
using a discrete cosine transformation with the number of coefficients set to
1/10 of the number of windows. The first ZCR value below the threshold set by
\code{zcr_min} is predicted to correspond to the voicing onset, unless there is
an adjacent gap in low ZCR values. See \link{vo_zcr}.
\item \code{soe}. The onset of voicing is predicted by passing a differenced,
downsampled version of the waveform through a cascade of zero frequency
filters to derive the strength of harmonic excitation (SoE), and choosing
the first zero crossing where the SoE is above a certain threshold set by
\code{soe_min}. See \link{vo_soe}.
}}

\item{vo_granularity}{Numeric, only used when \code{vo_method='acf'}. Mean energy
autocorrelation is calculated for intervals of this duration (in ms). Default
is \code{1}.}

\item{vo_param}{Numeric, only used when \code{vo_method='acf'}. Voicing onset
is predicted by comparing the mean autocorrelation in intervals of size
\code{vo_granularity} after the predicted closure
to the interval with the highest autocorrelation,
and finding the first interval where mean autocorrelation is at this
proportion of the most autocorrelated interval. Default is \code{0.85}.}

\item{f0_wl}{Numeric, only used when \code{vo_method='f0'}. The length of the
analysis window in ms passed on to \link[phonTools:pitchtrack]{phonTools::pitchtrack}. Default is \code{30},
much lower than the \link[phonTools:pitchtrack]{phonTools::pitchtrack} default of \code{50}, which seems
to give better, less conservative results when searching for the first pulses
after a stop burst.}

\item{f0_minacf}{Numeric, only used when \code{vo_method='f0'}. Autocorrelation
values below this are ignored by \link[phonTools:pitchtrack]{phonTools::pitchtrack}. Default is \code{0.5}.}

\item{zcr_min}{Numeric, only used when \code{vo_method='zcr'}. Voicing is
predicted to begin when ZCR values after the predicted release
are consistently below this threshold.
Default is \code{0.02}.}

\item{soe_min}{Numeric, only used when \code{vo_method='soe'}. Voicing is
predicted to begin by the first zero crossing with this SoE value in a
filtered version of the waveform. Default is \code{0.01}.}

\item{burst_only}{Logical; if \code{TRUE}, only burst location is predicted.}

\item{vo_only}{Boolean; default is \code{FALSE}. Can be set to \code{TRUE} if
\code{sign='positive'}, and the data is already aligned such that
intervals are aligned to the release. In this case, the burst location is
set as the beginning of the interval, and only voicing onset location is
predicted.}

\item{rel_offset}{Numeric, default is \code{0}. If \code{vo_only=TRUE}, the algorithm
may perform poorly if there is periodicity from e.g. voicing bleed early on
in the interval.
\code{rel_offset} tells \code{getVOT} how much of the initial portion of an interval
to ignore when looking for voicing onset (in seconds).}

\item{f0_first}{Logical; if \code{TRUE}, the function does not start out by
estimating the location of the closure, but rather uses \link[phonTools:pitchtrack]{phonTools::pitchtrack}
to find the longest period of consecutive voicing, and searches for the burst
in the 200ms chunk prior to the onset of that period.}

\item{plot}{Logical; should the results be plotted? Default is \code{TRUE}.}

\item{params_list}{A named list containing the above parameters for
determining the onset of voicing. Usually returned by \code{\link[=pos_setParams]{pos_setParams()}}, but
can also be written by hand. Should have the following named elements (order
not important; see details above):
\itemize{
\item \code{closure_interval}
\item \code{vo_granularity}
\item \code{vo_param}
\item \code{vo_method}
\item \code{f0_wl}
\item \code{f0_minacf}
\item \code{release_param}
\item \code{rel_method}
\item \code{f0_first}
If a \code{params_list} is provided, it will override any corresponding
parameters in the function call.
}}
}
\value{
A named list with the following elements:
\itemize{
\item \code{rel} Gives the sample number where the stop burst is predicted.
\item \code{vo} Gives the sample number where the onset of voicing is predicted.
\item \code{vot} Gives the predicted voice onset time in ms.
\item \code{spike_size} Gives the amplitude of the burst relative to the highest
measured amplitude in the signal. Potentially used by the \link{getVOT} function
to predict whether voice onset time is positive or negative.
}
}
\description{
Positive voice onset time and the location of pre-voicing is predicted for a
sound snippet containing a voiceless stop. Optionally, the wave is plotted
showing the predicted location of the stop release.
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs

positiveVOT(sound=sig, sr=sr)
}
\seealso{
This function is called by the more general and somewhat less
flexible function \link{getVOT}, which is in turn called by the functions
\link{VOT2newTG}, \link{addVOT2TG}, and \link{addVOT2emuDB} for bulk analyzing sounds
and annotating the output in new or existing TextGrids (for users of the
acoustic analysis software Praat) or an EMU database (for users of \link{emuR}).
The parameters can be be somewhat obscure and the default parameters do not
necessarily scale particularly well. If a small annotated data set
is provided, \link{pos_setParams} can be used to tune the parameters to your data
and annotation style.
}
