% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_amplitude.R
\name{rel_amplitude}
\alias{rel_amplitude}
\title{Predict location of stop release based on relative amplitude}
\usage{
rel_amplitude(sound, start, sq_rel, step, release_param = 15, smooth = FALSE)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave or differenced
speech signal.}

\item{start}{Numeric value giving the first sample of the search space in
the sound signal.}

\item{sq_rel}{Numeric vector determining the locations of chunks in the
sound snippet.}

\item{step}{The duration of each chunk in terms of samples (corresponding
to 1 ms).}

\item{release_param}{Numeric. The location of the burst is predicted by
searching for the first instance after the
estimated closure where a 1 ms interval has
an energy equal to or higher than this proportion of the maximum energy in
\code{sound}. Default is \code{15}.}

\item{smooth}{Boolean; should the release be estimated by searching for the
velocity peak in a smoothed time series of amplitude peaks? Default is
\code{FALSE}.}
}
\value{
When \code{smooth=FALSE}, a named list containing \code{rel}, the sample number
where the release is predicted to be, and \code{spike_size}, the size of the
amplitude peak at the predicted release. When \code{smooth=TRUE}, returns only
the sample number where the release is predicted to be.
}
\description{
Predicts the location of a stop release in a short chunked sound snippet
or differenced speech signal
by searching for which chunk is the first to have an amplitude peak above
a certain threshold relative to the highest amplitude in the snippet.
Optionally, the peaks are treated as a smoothed time series, in which case
the release is assumed to occur at the velocity peak of the smooth function.
The latter option is mainly used for estimating releases in pre-voiced stops.
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs
step <- 0.001*sr

clo <- find_closure(sig, sr)
sq_rel <- seq(from=clo, to=length(sig), by=step)
rel <- rel_amplitude(sig, clo, sq_rel, step)
}
\seealso{
This function is typically called by either \link{positiveVOT} or
\link{negativeVOT} with \code{rel_method='amplitude'}, \code{rel_method='raw'}, or
\code{rel_method='diff'}. These are in turn typically called when using
\link{VOT2newTG} or \link{addVOT2TG} to add predicted VOT to a Praat TextGrid or to an
EMU database with \link{addVOT2emuDB}.
}
