% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo_acf.R
\name{vo_acf}
\alias{vo_acf}
\title{Predict voicing onset in a stop based on autocorrelation}
\usage{
vo_acf(sound, start, end, baseline, step, granularity, param)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{start}{Numeric value giving the first sample of the search space in
the sound signal.}

\item{end}{Numeric value giving the final sample of the search space in
the sound signal.}

\item{baseline}{Numeric value giving the sample of a preceding landmark in
the sound signal (predicted closure when predicting negative VOT,
predicted release when predicting positive VOT).}

\item{step}{The duration of each chunk in terms of samples (corresponding
to 1 ms).}

\item{granularity}{Numeric. Mean energy
autocorrelation is calculated for intervals of this duration (in ms).}

\item{param}{Numeric. Voicing onset
is predicted by comparing the mean autocorrelation in intervals of size
\code{granularity} after the predicted closure
to the interval with the highest autocorrelation,
and finding the first interval where mean autocorrelation is at this
proportion of the most autocorrelated interval.}
}
\value{
A numeric value giving the sample number of predicted voicing onset.
}
\description{
Predicts the location of voicing onset in a short sound file from the mean
degree of energy autocorrelation in short intervals of the signal relative to
the mostautocorrelated interval during the vowel.
}
\examples{
datapath <- system.file('extdata/vd', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs
step <- 0.001*sr

clo <- find_closure(sig, sr)
vo <- vo_acf(sig, clo, length(sig), clo, step, 1.2, 0.9)
}
\seealso{
This function is typically called by either \link{positiveVOT} or
\link{negativeVOT} with \code{vo_method='acf'}.
These are in turn typically called when using
\link{VOT2newTG} or \link{addVOT2TG} to add predicted VOT to a Praat TextGrid or to an
EMU database with \link{addVOT2emuDB}.
}
