% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo_soe.R
\name{vo_soe}
\alias{vo_soe}
\title{Predict voicing onset in a stop based on strength of harmonic excitation}
\usage{
vo_soe(sound, start, sr, min_soe = 0.01)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{start}{Numeric value giving the first sample of the search space in
the sound signal.}

\item{sr}{Integer; sample rate of \code{sound}.}

\item{min_soe}{Numeric, only used when \code{vo_method='soe'}. Voicing is
predicted to begin by the first zero crossing with this SoE value in a
filtered version of the waveform. Default is \code{0.01}.}
}
\value{
A numeric value giving the sample number of predicted voicing onset.
}
\description{
Predict the location of voicing in a short sound file by passing a
differenced, downsampled version of the waveform through a cascade of zero
frequency filters to derive the strength of harmonic excitation (SoE), and
choosing the first zero crossing where the SoE is above a certain threshold.
}
\examples{
datapath <- system.file('extdata/vd', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs

clo <- find_closure(sig, sr)
vo <- vo_soe(sig, clo, sr)
}
\seealso{
This function is typically called by either \link{positiveVOT} or
\link{negativeVOT} with \code{vo_method='soe'}.
These are in turn typically called when using
\link{VOT2newTG} or \link{addVOT2TG} to add predicted VOT to a Praat TextGrid or to an
EMU database with \link{addVOT2emuDB}.
}
