% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vo_zcr.R
\name{vo_zcr}
\alias{vo_zcr}
\title{Predict voicing onset in a stop based on zero crossing rate}
\usage{
vo_zcr(sound, start, sr, minval, padNA)
}
\arguments{
\item{sound}{Numeric vector corresponding to a sound wave.}

\item{start}{Numeric value giving the first sample of the search space in
the sound signal.}

\item{sr}{Integer; sample rate of \code{sound}.}

\item{minval}{Voicing is
predicted to begin when ZCR values are consistently below this threshold.}

\item{padNA}{Boolean; should the first few ZCR values be treated as \code{NA} in
order to avoid predicted voicing onset immediately after the value set by
\code{start}? This is typically necessary when predicting positive VOT using ZCR.}
}
\value{
A numeric value giving the sample number of predicted voicing onset.
}
\description{
Predict the location of voicing in a short sound file by calculating ZCR in
5 ms windows with 80\% overlap, and smoothing the resulting time series
using a discrete cosine transformation with the number of coefficients set to
1/10 of the number of windows. The first ZCR value below a certain threshold
is predicted to correspond to the voicing onset, unless there is
an adjacent gap in low ZCR values.
}
\examples{
datapath <- system.file('extdata/vd', package='getVOT')
snd <- rPraat::snd.read(paste0(datapath, '/1.wav'))
sig <- snd$sig[,1]
sr <- snd$fs

clo <- find_closure(sig, sr)
vo <- vo_zcr(sig, clo, sr, 0.02, FALSE)
}
\seealso{
This function is typically called by either \link{positiveVOT} or
\link{negativeVOT} with \code{vo_method='zcr'}.
These are in turn typically called when using
\link{VOT2newTG} or \link{addVOT2TG} to add predicted VOT to a Praat TextGrid or to an
EMU database with \link{addVOT2emuDB}.
}
