% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VOT2newTG.R
\name{VOT2newTG}
\alias{VOT2newTG}
\title{Predict voice onset time and create TextGrid with the results}
\usage{
VOT2newTG(
  directory,
  verbose = TRUE,
  sign = c("positive", "negative"),
  neg_params_list = NULL,
  pos_params_list = NULL
)
}
\arguments{
\item{directory}{A string giving the location of a directory with WAV files.
The results will be stored in a new subdirectory called \code{tg}.}

\item{verbose}{Logical; should messages be printed in the console indicating
the progress? Default is \code{TRUE}.}

\item{sign}{A string giving the type of voice onset time to predict;
legal values are \code{positive}, \code{negative}, or both. Default is both, in which
case \link{getVOT} tries to predict whether voice onset time is positive or
negative. The development of this function is still in the early stages,
so it is recommended to specify whether VOT is expected to be positive or
negative when possible.}

\item{neg_params_list}{Named list of parameters used to predict negative
voice onset time; see \link{negativeVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}

\item{pos_params_list}{Named list of parameters used to predict positive
voice onset time; see \link{positiveVOT} for more information. Default is \code{NULL},
in which case default parameters are used, but note that the default
parameters do not necessarily scale particularly well.}
}
\description{
Loops through a directory with WAV files, each containing a syllable with
an initial stop consonant. Predicts voice onset time and generates a new
Praat TextGrid file giving the location of the predicted voice onset time
(i.e. pre-voicing or stop release).
}
\examples{
datapath <- system.file('extdata/vl', package='getVOT')
VOT2newTG(datapath)
}
\seealso{
WAV files are imported and TextGrids are exported using the \code{rPraat}
package. Voice onset time is predicted using the \link{getVOT} function, which
calls either the \link{positiveVOT} or \link{negativeVOT} function (or both) for a
given token. See also the sister function \link{addVOT2TG} which adds
predicted voice onset time to an existing TextGrid where the rough location
of a stop has been annotated.
}
